refactor: ProjectStateBadge를 /features/project로 이동
